refactor: consolidate main view into app.vue
